# contract/openapi.json
{"openapi":"3.1.0","info":{"title":"Web Respect consent receipt contract","version":"1.0.0","description":"Host-owned server-to-server API. UUIDs are correlation handles, not verified identity. Updates and withdrawals append records."},"servers":[{"url":"http://127.0.0.1:4330"}],"security":[{"gatewayToken":[]}],"paths":{"/v1/consents":{"post":{"summary":"Create/update/withdraw consent receipt","parameters":[{"$ref":"#/components/parameters/Subject"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"./consent.schema.json"}}}},"responses":{"201":{"description":"Receipt with id, server receivedAt and submitted record"},"422":{"description":"Invalid schema, policy, time or choices"},"401":{"description":"Unauthorized"},"403":{"description":"Origin rejected"},"413":{"description":"Body exceeds 8 KiB"},"429":{"description":"Rate limited"}}}},"/v1/consents/current":{"get":{"summary":"Latest choices and validity for host-authorized subject","parameters":[{"$ref":"#/components/parameters/Subject"}],"responses":{"200":{"description":"Receipt and valid boolean (policy and expiry)"},"404":{"description":"No record"},"401":{"description":"Unauthorized"}}}},"/v1/admin/consents":{"delete":{"summary":"Privileged privacy-request deletion for authorized subject","security":[{"adminToken":[]}],"parameters":[{"$ref":"#/components/parameters/Subject"}],"responses":{"200":{"description":"Deleted receipt count"},"401":{"description":"Privileged credential required"}}}},"/v1/admin/retention":{"post":{"summary":"Privileged configured audit retention cleanup","security":[{"adminToken":[]}],"parameters":[{"$ref":"#/components/parameters/Subject"}],"responses":{"200":{"description":"Deleted receipt count"},"401":{"description":"Privileged credential required"}}}}},"components":{"securitySchemes":{"gatewayToken":{"type":"http","scheme":"bearer","description":"Host gateway credential; never delivered to the browser"},"adminToken":{"type":"http","scheme":"bearer","description":"Distinct privileged worker credential"}},"parameters":{"Subject":{"in":"header","name":"X-WR-Subject","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,128}$"},"description":"Derived and authorized by the host, never trusted from browser input"}}}}